Remove 14 crate-level clippy allows by fixing violations
Fix code issues rather than suppressing clippy warnings:

- derivable_impls: Add #[derive(Default)] with #[default] on variants
- redundant_closure: Replace closures with unwrap_or_default()
- manual_div_ceil: Use .div_ceil() standard library method
- manual_clamp: Use .clamp() standard library method
- unnecessary_cast: Remove redundant type casts
- wrong_self_convention: Use self for Copy types, rename to_image→into_image
- field_reassign_with_default: Use struct init syntax with ..Default
- explicit_counter_loop: Use .enumerate()
- redundant_pattern_matching: Use while .is_some()
- non_canonical_clone_impl: Simplify clone impl
- empty_line_after_doc_comments: Fix doc comment placement
- ptr_arg: Change &mut Vec<u8> to &mut [u8] for slices

Add local allows with documentation for justified exceptions:
- module_inception: Platform-specific module pattern
- derivable_impls: ThemeBuilder uses semantic default
- upper_case_acronyms: GPU, IEEE, IQR are industry standard terms
- enum_variant_names: GpuError variants describe failure modes
- type_complexity: Complex types for GPU buffer cache and event handlers

Keep too_many_arguments crate-level with documentation (33 functions).

+++ src/render/gpu/renderer.rs
@@ -200,12 +200,10 @@ impl GpuRenderer {
         // Setup transformation parameters
         let (left, top, right, bottom) = viewport;
         let params = TransformParams {
-            scale_x: (right - left) as f32 / (x_range.1 - x_range.0) as f32,
-            scale_y: (bottom - top) as f32 / (y_range.1 - y_range.0) as f32,
-            offset_x: left
-                - (x_range.0 as f32 * (right - left) as f32 / (x_range.1 - x_range.0) as f32),
-            offset_y: top
-                - (y_range.0 as f32 * (bottom - top) as f32 / (y_range.1 - y_range.0) as f32),
+            scale_x: (right - left) / (x_range.1 - x_range.0) as f32,
+            scale_y: (bottom - top) / (y_range.1 - y_range.0) as f32,
+            offset_x: left - (x_range.0 as f32 * (right - left) / (x_range.1 - x_range.0) as f32),
+            offset_y: top - (y_range.0 as f32 * (bottom - top) / (y_range.1 - y_range.0) as f32),
             width: (right - left) as u32,
             height: (bottom - top) as u32,
             _padding: [0, 0],
